update sql queries, add enpoint in games

+++ services/games.js
@@ -6,15 +6,33 @@ const mysql = require("mysql2/promise");
 async function getMultiple(query) {
     const { offset, limit, q } = query;
 
-    const data = await db.query(
+    const games = await db.query(
         `SELECT g.id, g.name, g.price, IF(g.is_digital,(SELECT COUNT(gk.id) FROM games_keys gk WHERE gk.game_id=g.id),g.quantity) AS quantity,
          g.description, g.release_date, g.is_digital,
-         g.age_category, g.platform_id, JSON_ARRAYAGG(gc.category_id) as categories_id FROM games g, games_categories gc 
-         WHERE g.id = gc.game_id AND name LIKE ? GROUP BY g.id LIMIT ? OFFSET ?;`,
+         g.age_category, g.image_url, (SELECT JSON_OBJECT('id',p.id,'name',p.name) FROM platforms p WHERE p.id=g.platform_id) AS platform, 
+         JSON_ARRAYAGG((SELECT JSON_OBJECT('id',c.id,'name',c.name) FROM  categories c WHERE gc.category_id=c.id)) AS categories FROM games g, games_categories gc WHERE g.id=gc.game_id AND name LIKE ? GROUP BY g.id LIMIT ? OFFSET ?;`,
         [`%${q || ""}%`, limit || "1844674407370955161", offset || "0"]
     );
 
-    return data;
+    const [
+        total,
+    ] = await db.query(`SELECT COUNT(id) AS totalGamesNumber FROM games WHERE name LIKE ?`, [
+        `%${q || ""}%`,
+    ]);
+
+    return { totalGamesNumber: total.totalGamesNumber, games };
+}
+
+async function getOne(id) {
+    const game = await db.query(
+        `SELECT g.id, g.name, g.price, IF(g.is_digital,(SELECT COUNT(gk.id) FROM games_keys gk WHERE gk.game_id=g.id),g.quantity) AS quantity,
+    g.description, g.release_date, g.is_digital,
+    g.age_category, g.image_url, (SELECT JSON_OBJECT('id',p.id,'name',p.name) FROM platforms p WHERE p.id=g.platform_id) AS platform, 
+    JSON_ARRAYAGG((SELECT JSON_OBJECT('id',c.id,'name',c.name) FROM  categories c WHERE gc.category_id=c.id)) AS categories FROM games g, games_categories gc WHERE g.id=? AND g.id=gc.game_id GROUP BY g.id;`,
+        [id]
+    );
+
+    return game[0];
 }
 
 async function getMultipleSearch(query, params) {
@@ -22,9 +40,9 @@ async function getMultipleSearch(query, params) {
 
     const { name, is_digital, age_categories, platforms_id, categories_id } = params;
 
-    const data = await db.query(
+    const games = await db.query(
         `SELECT g.id, g.name, g.price, IF(g.is_digital,(SELECT COUNT(gk.id) FROM games_keys gk WHERE gk.game_id=g.id),g.quantity) AS quantity, g.description, g.release_date, g.is_digital,
-        g.age_category, g.platform_id, JSON_ARRAYAGG(gc.category_id) as categories_id FROM games g, games_categories gc 
+        g.age_category, g.image_url, (SELECT JSON_OBJECT('id',p.id,'name',p.name) FROM platforms p WHERE p.id=g.platform_id) AS platform, JSON_ARRAYAGG((SELECT JSON_OBJECT('id',c.id,'name',c.name) FROM  categories c WHERE gc.category_id=c.id)) AS categories FROM games g, games_categories gc 
         WHERE g.id = gc.game_id AND g.id=ANY(SELECT DISTINCT(g.id) FROM games g, games_categories gc WHERE g.id = gc.game_id 
         AND g.name LIKE ? ${
             is_digital && is_digital.length > 0
@@ -50,12 +68,34 @@ async function getMultipleSearch(query, params) {
         ]
     );
 
-    return data;
+    const [total] = await db.query(
+        `SELECT COUNT(DISTINCT(g.id)) AS totalGamesNumber FROM games g, games_categories gc WHERE g.id = gc.game_id 
+    AND g.name LIKE ? ${
+        is_digital && is_digital.length > 0
+            ? ` AND g.is_digital IN(${is_digital.map((i) => mysql.escape(i))}) `
+            : ``
+    } ${
+            age_categories && age_categories.length > 0
+                ? ` AND g.age_category IN(${age_categories.map((ag) => `${mysql.escape(ag)}`)}) `
+                : ``
+        } ${
+            platforms_id && platforms_id.length > 0
+                ? ` AND g.platform_id IN(${platforms_id.map((id) => mysql.escape(id))}) `
+                : ``
+        } ${
+            categories_id && categories_id.length > 0
+                ? ` AND gc.category_id IN(${categories_id.map((id) => mysql.escape(id))}) `
+                : ``
+        };`,
+        [name && name.length > 0 ? `%${name}%` : "%%"]
+    );
+
+    return { totalGamesNumber: total.totalGamesNumber, games };
 }
 
 async function create(game) {
     let queries = [
-        `INSERT INTO games (name, price, quantity, description, release_date, is_digital, age_category, platform_id) VALUES (?,?,?,?,?,?,?,?)`,
+        `INSERT INTO games (name, price, quantity, description, release_date, is_digital, age_category, platform_id, image_url) VALUES (?,?,?,?,?,?,?,?,?)`,
     ];
     let queryValues = [
         [
@@ -67,6 +107,7 @@ async function create(game) {
             game.is_digital,
             game.age_category,
             game.platform_id,
+            game.image_url,
         ],
     ];
     game.categories_id.map((c) => {
@@ -99,6 +140,7 @@ async function updateChosen(id, game) {
                 "is_digital",
                 "age_category",
                 "platform_id",
+                "image_url",
             ],
             [
                 game.name,
@@ -109,20 +151,21 @@ async function updateChosen(id, game) {
                 game.is_digital,
                 game.age_category,
                 game.platform_id,
+                game.image_url,
             ]
         )} WHERE id=?`,
     ];
     let queryValues = [[id]];
 
-    if (game.categories_id) {
+    if (game.categories_id && game.categories_id.length > 0) {
         queries = [...queries, `DELETE FROM games_categories WHERE game_id=?`];
         queryValues = [...queryValues, [id]];
         game.categories_id.map((c) => {
             queries = [
                 ...queries,
-                `INSERT INTO games_categories (game_id, category_id) VALUES ((SELECT MAX(id) FROM games),?)`,
+                `INSERT INTO games_categories (game_id, category_id) VALUES (?,?)`,
             ];
-            queryValues = [...queryValues, [c]];
+            queryValues = [...queryValues, [id, c]];
         });
     }
     const [result] = await db.transaction(queries, queryValues);
@@ -158,6 +201,7 @@ async function remove(id) {
 
 module.exports = {
     getMultiple,
+    getOne,
     getMultipleSearch,
     create,
     updateChosen,
